Reframe the API towards the atomic job
Previously the API assumed the 1:1 mapping between work (executables)
and a target which was suitable for UT-style testing where workers
cannot or are not supposed to affect each other.

Point of this change is to enable the parallel, somewhat synchronized
execution of possibly different executables on different targets and
capturing data from all of them upon completion.